% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crossvalid}
\alias{crossvalid}
\title{Cross validation routine for species distribution models}
\usage{
crossvalid(Ncv, Kfold, data, method, responsetype, response, predictors,
  secondary = NULL, strata, enviStack, enviPix, seed, check.names = TRUE,
  aggregated = NULL, pseudoabsence = NULL, gbm.trees = 2000,
  maxargs = c("outputformat=logistic", "defaultprevalence=0.5"),
  rast = TRUE, flat = FALSE, ...)
}
\description{
run cross-validation
}

