% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model}
\alias{model}
\title{Make a species distribution model}
\usage{
model(data, method, responsetype, response, predictors, secondary = NULL,
  enviStack, enviPix, seed = NULL, aggregated = NULL,
  pseudoabsence = NULL, gbm.trees = 2000,
  maxargs = c("outputformat=logistic", "defaultprevalence=0.5"),
  model = FALSE, prediction = FALSE, rast = TRUE, rast.occ = TRUE,
  flat = FALSE, moran = FALSE, ...)
}
\arguments{
\item{data}{\code{SpatialPointsDataFrame} containing response and predictors}

\item{method}{SDM methos used: "gam", "rf", "gbm", "max", or "gbm.step". See details for details}

\item{responsetype}{Type of the response: "count", "continuous", or "presence". Presence denotes a bimodal response 0|1}

\item{response}{}

\item{predictors}{}

\item{secondary}{}

\item{enviStack}{}

\item{enviPix}{}

\item{seed}{}

\item{aggregated}{}

\item{pseudoabsence}{}

\item{gbm.trees}{}

\item{maxargs}{}

\item{model}{}

\item{prediction}{}

\item{rast.occ}{}

\item{flat}{}

\item{moran}{}

\item{...(ellipsis)}{}
}
\value{
A \code{model}, or a \code{data.frame} or a \code{list} depending on arguments
}
\description{
Allocate an empthy data.frame from template dataframe
}
\details{
ysdgadghadfgh
}
\examples{
Notyet
}

