% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{quansdm}
\alias{quansdm}
\alias{quansdm-package}
\title{quansdm: quantitative species distribution modelling framework}
\description{
quansdm provides wrapper functions for some popular species distribution modelling methods.
}
\details{
It has three main goals:
\itemize{
\item Free you from tedious tasks.
\item Provide a large variety of model performance metrics.
\item Provide model validation techniques like cross-validation, null-models,
and  exploration of spatial autocorrelation.
}
}

