% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crossvalid}
\alias{crossvalid}
\title{Cross validation routine for species distribution models}
\usage{
crossvalid(Ncv, Kfold, data, method, responsetype, response, predictors,
  secondary = NULL, strata, enviStack, enviPix, seed, check.names = TRUE,
  aggregated = NULL, pseudoabsence = NULL, gbm.trees = 2000,
  maxargs = c("outputformat=logistic", "defaultprevalence=0.5"),
  rast = TRUE, flat = FALSE, ...)
}
\arguments{
\item{Ncv}{Number of cv repititions.}

\item{Kfold}{Number of folds. Usually take 10 or 5.}

\item{rast}{return \code{list} of metric and raster}

\item{flat}{return model performance metrics only (as \code{data.frame})}

\item{...}{ellipsis is used to pass arguments to subsequent functions like  \code{threshold.def}. See \code{\link{metrics}} for details}
}
\description{
repeated k-fold cross-validation. Calculate model preformance metrics, disribution, standard deviation and occurence maps.
}

