% Generated by roxygen2 (4.0.2): do not edit by hand
\name{model}
\alias{model}
\title{Make a species distribution model}
\usage{
model(data, method, responsetype, response, predictors, secondary = NULL,
  enviStack, enviPix, seed = NULL, aggregated = NULL,
  pseudoabsence = NULL, gbm.trees = 2000,
  maxargs = c("outputformat=logistic", "defaultprevalence=0.5"),
  model = FALSE, prediction = FALSE, rast = TRUE, rast.occ = TRUE,
  flat = FALSE, moran = FALSE, ...)
}
\arguments{
\item{data}{\code{SpatialPointsDataFrame} containing response and predictors}

\item{method}{SDM methos used: "gam", "rf", "gbm", "max", or "gbm.step". See details for details}

\item{responsetype}{Type of the response: "count", "continuous", or "presence". Presence denotes bimodal responses [0|1]}

\item{response}{Column name of response in data argument}

\item{predictors}{Column names of predictors to use in data argument}

\item{secondary}{Column name in data. Use this to calculate model performance metrics from instead of response. Default is NULL.}

\item{enviStack}{\code{RasterStack} of predictors. Used to calculate SD map}

\item{enviPix}{\code{SpatialPixelsDataFrame} of predictors. enviPix<-as(enviStack,"SpatialPixelsDataFrame"). Only for performance.}

\item{seed}{Integer. For reproduceability.}

\item{aggregated}{logical or NULL. Default is NULL.}

\item{pseudoabsence}{logical or NULL. Default is NULL}

\item{gbm.trees}{gbm.trees param for dismo::gbm}

\item{maxargs}{argument to pass tp maxent}

\item{model}{return only the model. Overrides args below}

\item{prediction}{return only the prediction. Overrides args below}

\item{rast}{return \code{list} of metric and raster}

\item{rast.occ}{additionally adds bimodal occurence map. see details for threshold calculation.}

\item{flat}{return model performance metrics only (as \code{data.frame})}

\item{moran}{}

\item{...(ellipsis)}{threshold.def}
}
\value{
A \code{model}, or a \code{data.frame} or a \code{list} depending on arguments
}
\description{
make a species distribution model
}
\details{
ysdgadghadfgh
}
\examples{
Notyet
}

